moved components to be nested under navlink
