+++ frontend/src/components/Navigation/index.js
@@ -1,39 +1,68 @@
+// import React from 'react';
+// import { NavLink } from 'react-router-dom';
+// import { useSelector } from 'react-redux';
+// import ProfileButton from './ProfileButton';
+// import OpenModalButton from '../OpenModalButton';
+// import LoginFormModal from '../LoginFormModal';
+// import SignupFormModal from '../SignupFormModal';
+// import './Navigation.css';
+
+// function Navigation({ isLoaded }){
+//   const sessionUser = useSelector(state => state.session.user);
+
+//   let sessionLinks;
+//   if (sessionUser) {
+//     sessionLinks = (
+//       <li>
+//         <ProfileButton user={sessionUser} />
+//       </li>
+//     );
+//   } else {
+//     sessionLinks = (
+//       <li>
+//         <OpenModalButton
+//           buttonText="Log In"
+//           modalComponent={<LoginFormModal />}
+//         />
+//         <OpenModalButton
+//           buttonText="Sign Up"
+//           modalComponent={<SignupFormModal />}
+//         />
+//       </li>
+//     );
+//   }
+
+//   return (
+//     <ul>
+//       <li>
+//         <NavLink exact to="/">Home</NavLink>
+//       </li>
+//       {isLoaded && sessionLinks}
+//     </ul>
+//   );
+// }
+
+// export default Navigation;
+
 import React from 'react';
 import { NavLink } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 import ProfileButton from './ProfileButton';
-import OpenModalButton from '../OpenModalButton';
-import LoginFormModal from '../LoginFormModal';
 import './Navigation.css';
 
 function Navigation({ isLoaded }){
   const sessionUser = useSelector(state => state.session.user);
-
-  let sessionLinks;
-  if (sessionUser) {
-    sessionLinks = (
-      <li>
-        <ProfileButton user={sessionUser} />
-      </li>
-    );
-  } else {
-    sessionLinks = (
-      <li>
-        <OpenModalButton
-          buttonText="Log In"
-          modalComponent={<LoginFormModal />}
-        />
-        <NavLink to="/signup">Sign Up</NavLink>
-      </li>
-    );
-  }
 
   return (
     <ul>
       <li>
         <NavLink exact to="/">Home</NavLink>
       </li>
-      {isLoaded && sessionLinks}
+      {isLoaded && (
+        <li>
+          <ProfileButton user={sessionUser} />
+        </li>
+      )}
     </ul>
   );
 }
